Подтягивание данных в новую форму, добавлена обработка предметов и выставвление оценок к ним. Нужно настроить сохранение состояния оценок предметов за студентом

+++ Models/Subject.cs
@@ -9,7 +9,8 @@ namespace WFA
 	public class Subject
 	{
 		public string Name { get; }
-		public byte Mark;
+
+		public int Mark;
 
 		public Subject(string name)
 		{
